Add tests for EVM portable signatures
EVM portable signatures are particularly suitable for the protection of
metadata of immutable files where metadata is signed by a software vendor.
They can be used for example in conjunction with an IMA policy that
appraises only executed and memory mapped files.

However, some usability issues are still unsolved, especially when EVM is
used without loading an HMAC key. The kernel patch set 'evm: Improve
usability of portable signatures' attempts to fix the open issues.

The purpose of the new tests is to verify that operations on files with EVM
portable signatures succeed and that the new kernel patch set does not
break the existing kernel integrity expectations.

To run the tests, it is necessary to pass the path of the kernel private
key with the TST_KEY_PATH environment variable. If not provided, the script
searches the key in /lib/modules/$(uname -r)/source/certs/signing_key.pem
and in the current directory. Root privileges are required to mount the
image, configure IMA/EVM and set xattrs.

The script attempts to launch a UML kernel named 'linux' if it is found in
the ima-evm-utils root directory.

Only the tests with a compatible EVM mode will be executed, unless the
TST_EVM_CHANGE_MODE variable is set to 1 (if the tests are executed by a
UML kernel TST_EVM_CHANGE_MODE is always set to 1). In this case, the
script attempts to change the current EVM mode.

The check_evm_revalidate() test can be executed only if
EVM_ALLOW_METADATA_WRITES is set in advance before running the tests. To do
it, it is sufficient to execute:

echo 4 > /sys/kernel/security/evm

This step is not necessary if the script launches a UML kernel.

Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>