Download mount-idmapped
mount-idmapped is a tool to create idmapped mounts, a feature recently
integrated in kernel 5.12 and that allows processes to see a different UID
and GID on that mount.

This patch downloads (errors are ignored) an artifact from
$MOUNT_IDMAPPED_URL/mount-idmapped, where MOUNT_IDMAPPED_URL is an
environment variable to be set in the configuration of the testing platform
(for Github Actions, the variable should be added in an environment named
'test').

Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>