Add additional options to the container engine
This patch adds additional options to the command line of the container
engine, necessary to run the UML kernel.

The additional options are: --cap-add=SYS_PTRACE -v /dev/shm:/dev/shm

The first adds the PTRACE capability to the container, and the second adds
a bind mount for /dev/shm.

Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>