Add support for vm environment
Add the new 'vm' driver, to start a virtual machine with QEMU and run the
provided kernel. If swtpm is present in the system, also instantiate a TPM
emulator and attach it to the virtual machine.

The 'vm' driver has a similar behavior of the 'um' driver, with the
exception that the built kernel needs the 9p filesystem instead of hostfs
(UML-specific) to have visibility of the host filesystem and relaunch the
script in the new environment.

The virtual machine architecture (suffix of the qemu-system- binary) can be
specified with the TST_ARCH environment variable.

Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>