Add support for UML in functions.sh
Add the new functions _run_user_mode(), _exit_user_mode(),
_init_user_mode() and _cleanup_user_mode() to run the tests inside a system
booted with the UML kernel.

A typical structure of a script with tests is:

trap cleanup SIGINT SIGTERM SIGSEGV EXIT

_cleanup() {
	<test cleanup>
}

cleanup() {
	_cleanup_user_mode _cleanup
	_report_exit_and_cleanup
}

<tests implementations>

_run_user_mode ../linux $PWD/$(basename $0) "env_var1=$env_var1 ..."

_exit_user_mode ../linux

_init_user_mode

<tests init>

<tests call>

If the UML_MODE environment variable is not set to 1, or the UML kernel
does not exist, ignore the UML kernel execution and initialization
requests, and perform the cleanup in the current environment. Ignore the
same also if the script is already run in the UML enviroment, to avoid
loops. Instead, for cleanup, do it only in the UML environment and skip it
in the host environment.

Signal to the host environment failures of tests run in the UML environment
with an unclean shutdown of the UML kernel.

Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>