Add kernel configuration for tests
Add kernel-configs/base with changes to be applied to the default kernel
configuration, generated with 'make defconfig'.

Add kernel-configs/integrity, with integrity-specific configuration
options.

Splitting changes helps to identify more easily the desired group of
options. In the future, options could be split even further.

All changes in this directory will be applied with the merge_config.sh
script from the kernel source code in a Github workflow step.

Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>