Adapt fsverity.test to work with UML kernel
Adapt fsverity.test by adding calls to the UML kernel API in
functions.sh. If the UML_MODE environment variable is set to 1, and the
UML kernel exists in the ima-evm-utils directory, run the UML kernel
first and execute the tests in the new environment. Otherwise, keep the
current behavior.

Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>