Pass cleanup function to _report_exit_and_cleanup()
If an error occurs before any test is executed, for example during
initialization, this error is not returned to the shell as the exit code
currently depends on the number of tests succeeded/skipped/failed.

Pass the cleanup function and its argument to _report_exit_and_cleanup(),
so that the latter can get the current exit code at script exit or when a
signal is received and can return that exit code to the shell if no test
was executed.

Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>