Add ima_policy_check.awk and ima_policy_check.test
Add the awk script to check for possible interference of a rule added by a
test with the existing IMA policy (policy replacement at the first policy
load is not taken into account).

The script expects the rule to be added to be the first, and the existing
IMA policy to be after.

Rules are considered as interfering if they don't overlap, or they overlap
but they have the same action and same policy options with the same values.

Combinations of MMAP_CHECK and MMAP_CHECK_REQPROT hooks are always
considered as interfering, due to the different behavior of those hooks.

Finally add a test, to ensure that the awk script behaves as expected.

Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>