Add simple test to check EVM HMAC calculation
Add a simple test to ensure that the kernel and evmctl provide the same
result for the HMAC calculation. Also add another test to evaluate the
HMAC on a directory for which Smack added the SMACK64TRANSMUTE xattr.

The tests require that, unless the UML kernel is used, the
TST_EVM_CHANGE_MODE environment variable is set to 1.

Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>