Add simple test to check EVM HMAC calculation
Add a simple test to ensure that the kernel and evmctl provide the same
result for the HMAC calculation. Also add another test to evaluate the
HMAC on a directory for which Smack added the SMACK64TRANSMUTE xattr.

While it was expected to see the second test failing without the patch
'smack: Set the SMACK64TRANSMUTE xattr in smack_inode_init_security()',
that in reality didn't happen due to EVM not checking the HMAC of
directories. The Smack test could be done only on them.

Require (unless the UML kernel is used) that the TST_EVM_CHANGE_MODE
environment variable is set to 1, so that users acknoledge that they are
initializing EVM with a well-known HMAC key, which can introduce obvious
security concerns.

Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>